ENRESCUE stops returns and throws by default, /RELAX permits
ENRESCUE is typically used to sandbox code, and if you THROW or RETURN past
it then you don't get a good sandbox.

This makes the default error in those cases, but has /RELAX to allow
non-erroring throws.

It may be the wrong default, and it should be ENRESCUE/STRICT that stops
the throws, but I don't think so.

Also changes to be able to give back NIHIL vs decay to void, so maximum
information by default.  Again this may not be the best default.

Plus kills off RESCUE+, because it's likely a bad idea (and antiform
objects will likely be killed off soon)