Make Console use libRebol, Interruptibility on Level
This transitions the C portion of the console code to be based entirely
on libRebol.

In order to do that, it can't fiddle flags regarding interruptbility,
and then fiddle them back when something like a continuation is over.
Reflecting on the issue, it seems that interruptibility should be
a property on the Level, and by doing so you don't have to fiddle
the flags--the state is just in sync with the level--so if you drop
an interruptible level you return to the interruptibility of the
parent.

This causes many places that were not acting interruptible before to
be so, exposing some hidden bugs.  Those bugs will be tackled in a
later commit.