Rename SET_SIGNAL() etc. to Set_Trampoline_Flag()
Rebol had a concept of requests you could make to the interpreter
that it would heed in-between evaluations.  It confusingly called
these flags "signals" and used functions like SET_SIGNAL() with
"sigmask", which conflates with unix signals.

This just pushes them to being TRAMPOLINE_FLAG_XXX, because the
trampoline in this case is acting kind of like the "OS" in terms
of getting control to do the "signaly" things when a native has
delegated to it (by returning or running a continuation).

Also, to avoid having to deal with the signal masking, the top
level is now encoding interruptibility.  This means that when the
level is dropped off the stack, there's no need to fiddle with
flags  So state saving and restoring doesn't have to worry about
the flags.