Eliminate SET-PATH! and GET-PATH!
As a prelude to migrating SET-XXX! and GET-XXX! to be built on top of
a CHAIN! sequence type, this gets rid of all SET-PATH! and GET-PATH!
instances.

Beyond getting rid of demonstrations that the part could exist, the
only real usage was when :APPEND/DUP was used to make a specialization.
Those instances are changed to GET $APPEND/DUP.

A result of this is that this Ren-C cannot bootstrap itself until the
bootstrap sources are changed to use modern notation.  That will happen
when the bootstrap executable is modified to load the modern notation
as if it were a PATH! for its own evaluation.