Allow obj.field selection syntax (converted to PATH!)
This does some trickery in order to make obj.field scan as obj/field

It doesn't offer any benefits in this build besides readability.  But it
allows the bootstrap code that uses this executable that is shared with the
modern executable to enforce the rule that slashes only run actions.