Adopt modern term Bounce for return type from natives
"REB_R" was used to mean "Rebol Return Value" from a native in the
old codebase.  Today this is a "Bounce", which was named that because
what you're returning is being given to the trampoline and can be
a continuation.

While there are no continuations, the idea of calling it "Bounce"
makes sense as it can be a value or a signal which might be a throw.
In any case, consistency is good for anyone unlucky enough to be
trying to read the old code and map it to the new.