Change over to dot selection of members in methods
The modern idea in Ren-C for methods in derived classes being able to access
the fields of the derived object (instead of being stuck with the bindings
to the original object) isn't to copy and rebind the methods on derivation
(as in historical Redbol) or to use "derived binding" (as in historical
Ren-C).  Instead, when an object method is invoked e.g. `obj/accessor` the
path selection creates a FRAME! cell with the object embedded in it (what
is called a "coupling").  Then when blank headed tuples--such as .field--
are used, the lookup of the term happens in the coupling.

This changes Rebmake to use that notation.  The bootstrap executable does
not know what to make of it, so it just loads `.abc` as `abc`.