Catch initializations of ~null~ into API handles
This puts a check in the core routine for initializing words that
helps catch locations that put nulls into API cells.

The currency of null as far as the user is concerned in the API
is supposed to be the language's null (C null, JavaScript null...)
in order to be falsey.  If a pointer to an allocated API handle
is initialized with null cell bits, that would not appear falsey
to the API user's code.