"Flexible Logic": TRUE and FALSE are words
This removes the idea of trying to have a ~true~ and ~false~ (or `#[true]`
and `#[false]`) concept, instead relying on nullness to signal when not to
take a branch...and using WORD! as the state of logic variables.

It does mean you have to include the right boolean test for whatever words
you happen to use (e.g. IF TRUE? VAR or IF OFF? VAR).  But it encourages
more freedom to choose, and avoids historically thorny issues of conflating
the logic state with the word--when the words are what people want to see
in the interface and exchange as currency.

  https://forum.rebol.info/t/flexible-logic-system-terminology/2252