Primordial GETTER/SETTER (ACCESSOR) and SET/GET cleanup
The idea has been on the table for some time that variable cells would
carry a special bit, saying that they hold an action which is not
meant as the content of the cell...but an action which is intended
to be run to get or set the actual variable.

This creates a lot of potential for code that wasn't written with an
understanding of accessors to misinterpret the contents of a variable
and either overwrite the accessor or treat it as if it were the
intended value to fetch.  To help catch these cases, the existing
bits for protecting cells against reading and writing are set on
the accessor cells, and only accessor-aware code will subvert that
when it sees the accessor bit is set.

A consequence of this is that almost all code which does getting and
setting of variables will need to go through higher-level routines.
There was sort of a hodgepodge of mixed behaviors in the existing
code...making it not clear when it would handle voids and give
back nulls, or handle the case of the "variable" being a GROUP!
and not a TUPLE! or WORD! or PATH!.  This does some preliminary
cleaning of that, in anticipation of rewriting more of the core to
use accessor-aware functions vs. getting variable slots directly.