Get POINTFREE Working Again
It hadn't been a priority for pointfree to work right now, but this
bites the bullet and fixes the demo.

Mostly just cleanup, but this involved changing RUNS so that it
would "harden" the frame into something that can actually be used
as a function that gathers arguments at a callsite.  Current plain
FRAME!s cannot do this correctly (the design is under review to
be able to permit it.)