Invalid picks cause definitional error
When TUPLE! could dispatch action, there was a risk of conflating
a definitional error from a bad pick with a definitional error that
was returned by the function.  Now that tuples do not run functions,
the issue is gone... and we can give an error you can defuse with
TRY on bad picks.

    >> block: [a b c]

    >> block.4
    ** Error: bad pick

    >> try block.4
    == ~null~  ; anti