Combinators literal match first, type match second
There was an idea that making generic combinators like TAG! dispatch to
the literal tag instances in the combinator map would provide a place
to put hooks that could then apply to and override all tags.  This
was misguided, in the sense that the better way to get control to
override combinators is to not use a map with those combinators
in it, but instead delegate to them via a sub-map of some kind.  Doing
otherwise is a performance drain that also requires the generic
combinator to speak all possible protocols that the overridden elements
would speak.  That was seen in trying to deal with adding a new property
of combinators being /NEGATED, and a useless TAG! combinator struggling
with tunneling that intent.