Improve EVARS Leak Detection
There was an implementation trick to avoid silently leaking EVARS to piggy
back on the leak detection of stubs.  But the way it worked, you wouldn't
find out about the leak until shutdown...which was far after the leak, as
well as potentially too late to get information about where that leak came
from, and when the system was in no state to be able to run things like a
Probe().

This commit goes ahead and adds a global count of how many evars are
outstanding so that it can be caught in the moment.  Also fixes the leak
that triggered needing to do this change.