Add in concept of non-abrupt SYS.UTIL/EXIT shutdown
It will be more desirable for people to solve some problems by using
SYS.UTIL/EXIT when they can't get ahold of a definitional QUIT.  But
calling `exit()` without going through a shutdown procedure will leak,
and make it hard to know if there were problems when using address
sanitizer or Valgrind.

Making the trampoline cooperatively exit is a large work item.  This
adds an /ABRUPT refinement as a placeholder to act as a reminder that
work should be done at some point.