TRIPWIRE=antiform tag, HOLE=antiform parameter, VACANCY?
This introduces the concept of VACANCY? as the category which subsumes
antiform tag, antiform parameter, and nothing.  They all cause errors
when accessed via WORD! without special handling.  And they are all
considered values that are overridden by DEFAULT.

It also finally names antiform tag to be TRIPWIRE and antiform
parameter to be HOLE.