Kill off PARAMCLASS_OUTPUT, typecheck packs
Multi-returns started out prior to the existence of antiforms, and were
rigged up to set up relationships between variables in a function's frame
and a SET-BLOCK! on the left-hand side.

As they evolved and block antiforms became parameter packs, it became less
of a win to have the core be involved in proxying local variables into
parameter pack items.  It introduced a ton of nebulous questions that were
worse than the existing questions about exposing RETURN as an interface
variable.

It's fair to say that the core would have never gotten involved if pack
antiforms had existed, but instead would have just typechecked packs.
Any function generator that wanted to proxy locals to be implicit
return results would do that on its own dime.

This scraps the output parameters, type checks packs, and frees up @
parameters to be taken from callsites with binding