Patch WRITE-STDOUT to work in VSCode terminal
Not sure why, but writing large amounts of data to the VSCode terminal
using whatever method the device layer goes by causes it to print a lot
of blank space.  Printing in smaller chunks seems to work.  Not worth
investigating the old device model code at this time for an edge case
that only reproduces in the VSCode piping, so print by chunks in
WRITE-STDOUT (despite the fact that the insides of WRITE-STDOUT are
already chunked...)