Approximate definitional errors by making ERROR! ornery
This makes ERROR! not work in more contexts (like REDUCE, COMPOSE,
or appending to blocks).  It's not the right answer to not let you
work with errors as reified objects--but the bootstrap executable
has more to gain from acting like the modern executable w.r.t.
definitional errors than it has to gain from putting them in blocks.