Make LIT-WORD! equivalent to WORD! in FOR-EACH, etc.
In the modern interpreter we want FOR-EACH variables to be quoted
at the callsite, vs. implicitly taken literally.  This has been
decided as worth it to have the "litter" of the tick mark in order
to more clearly denote that it's not an already existed and bound
variable.

Rather than retrofit the bootstrap executable to this convention,
this instead just makes the literal quote accept seeing a LIT-WORD!
at the callsite.  This means that the old convention of saying a
lit-word meant don't create a variable is no longer applicable
(it was a bad choice, but there wasn't much of other choices in
the olden days).  Modern interpreter uses @word to say reuse the
binding, but that's not available.  Use #word instead, since that
is an ISSUE!