Stop using ATTEMPT in bootstrap
The bootstrap executable has renamed its trapping abstractions
to where it's SYS.UTIL/RESCUE, and TRANSCODE will return an
error! if it can't process the data.

It also approximates definitional errors by not letting you put
errors in blocks (it's not that necessary).

A better poor man's error strategy may be coming to the bootstrap
executable (e.g. a RAISED! datatype)