Don't double close handles in Windows Call
If writing to a file, there was some code that was closing fields of
the STARTUPINFO structure in the case of using a FILE! for output or
for input.  However, the design (which I think I made :-/) seems to
suggest that you shouldn't close those "sink" handles, you should
close the "write" and "read" handles instead which were set up
as being distinct if closing was required.  This appears to fix
redirecting to files in Windows (it was crashing).