Permit raised errors in packs, use in EVAL
This switches from the policy of not allowing raised errors in
antiform blocks to: allow them, but don't have PACK make them
by default.  You have to use PACK*

A raised error in a pack won't be discarded on decay, but will
fail...giving some amount of safety to being willing to use
the construct.