Rethink DELIMIT some, # is space
This hacks in the idea of BLOCK! tight forming in DELIMIT, but
only one level deep of that.  It also migrates # to being the
more useful space character (NUL not useful, and the reasons
for picking it to be NUL to be the "mostly-meaningless truthy
type" are now all overridden by ~okay~, the antiform with no
meaning beyond branch-trigger).

It also brings in @ as a molding operation in DELIMIT, as well
as letting you PRINT any arbitrary value with `print @x`.
This idea was done with quoting and `print ^x` at one time, but
now that @ exists it's a better choice.