Rework HELP into dialected and non-dialected parts
This shortens the information given back from just plain `help` to
not include description of how to use the HELP command itself...
suggesting you instead do HELP HELP for those details.

Those details explain the dialected behaviors of HELP, as well
as pointing you to how to subvert that by passing HELP a GROUP!
to evaluate.  Then the main HELP command focuses on the switch
statement implementing the dialected behavior, while dumping
individual values is broken out into a separate routine.

The resulting code is a lot clearer, though it's really just a
jumping off point for some "help czar" to go through and work
out in a more comprehensive way.  (But it's good to see that a
lot of the shaky code can be replaced by more confident code.)