Improve ~unreadable~, eliminate FLAVOR_MIN_TRASH_OK
This transitions the unreadable element state to being more than just
a debug feature.  It allows the unreadable bit to be taken advantage
of by MAP! (for instance), where it's not necessary to go to using
antiforms to get out of band with array elements...and also, to get
the advantages of asserts on uninformed access of the slots.

This lets us get rid of the idea of "arrays in which most antiforms
aren't legal, just trash" which was a poor approximation of being
able to have unreadables where the flag persists in release builds.