Fix atrophy of C build (overloaded Do_Array_At())
ANY-ARRAY! was renamed to ANY-LIST! in order to distinguish the
internal storage type `Array` from the cell category (that includes
an index and binding):

  https://forum.rebol.info/t/any-array-vs-any-list/1384

Excising the name "Any_Array" from functions led to renaming
Do_Any_Array_At_Throws() to just Do_At_Throws().  This created an
accidental reuse of the name Do_At_Throws() so it was a duplicate
symbol.  It did not cause errors in C++ builds because they were
unique overloadings, but would break C builds.

This fixes that, and adjusts the terminology a bit (DO is no
longer used to refer to the mere act of evaluating lists.)