Disallow BIND to WORD! as bind context synonym
R3-Alpha didn't have the ability to pick out a single word from a
context and bind it in some scope--it was chosen to be all or
nothing (for some reason).  Hence, BIND to a WORD! was taken to
mean you wanted to bind to the context of that word.

Being able to pick one word to bind to is important in new Ren-C,
so this gets rid of that idea.  If you want to bind to the context
of a word, you have to say that explicitly:

So not:

    bind block word

But:

    bind block binding of word

It would be technically possible to retrofit this to bind mutably
to single words, but not necessary for bootstrap at this time.