R3C's bootstrap shims 8994d23 CALL to default /WAIT
The original behavior of CALL was to not assume you wanted to
synchronously wait for a response.  So it would spawn a process,
give you back a PID, and return control to the interpreter.

This is not the most useful behavior most of the time, so it was
changed to be synchronous by default.  However the bootstrap shim
did adapt the older 8994d23 build to this policy.  Hence if you
tried to build R3C without a makefile and using the CALL facility,
it would run the compilation and generation of prep files in
parallel, leading to things flying by rather quickly and a lot
of build errors.