Make GROUP! the only soft escapable type
Experience bore out that making GET-GROUP! the escapable type for
literals instead of GROUP! was a mistake--and it was better to use
plain GROUP! as in historical Redbol.  The leading colon made it
consistent with allowing GET-WORD! and GET-TUPLE! to escape...but
it was just consistently ugly.

In fact, being able to escape colon-leading types is so rarely of
interest that this ditches it entirely (especially since colons are
not going to have anything to do with "getting" any more).