Some tidying of legacy non-stackless Eval code
While most all native code has been converted to stackless, there are some
stragglers that are waiting pending review and rewrite.  This notices that
a couple of the service routines for stackful invocations of the
evaluator aren't used any more (or are used just once, so can be folded
into the callsite to discourage more usages).

Also shifts away from using the term "Do".  It was once the case that EVAL
referred to a single step while "DO" meant run an evaluation to the end,
but this is no longer the case.  DO is a higher level operation that does
not run blocks.