Improvements like IS_VARLIST() => Is_Stub_Varlist()
This replaces the all-caps functions for testing the flavors of
stubs with a more informative name that reinforces what is testing
is a stub.

It also clears up some ambiguity about the category of "string
stubs".  Previously there was FLAVOR_SYMBOL and FLAVOR_STRING, which
is confusing because Symbol is a subclass of String.  This reworks
the idea that the stub superclass is FLAVOR_UTF8 to saying that
string is the superclass, with members SYMBOL and NONSYMBOL.