Rework how errors are returned in sequence init
When the original implementation of sequence initialization was
being worked out, it tried to give back a hint of what caused the
init to fail by putting that thing in the same output slot where
a valid sequence result would be...and then a null return result
would indicate that the cell was indicative of an error.  This
is a narrow and confusing interface when all that's going to
happen is an error be reported...so returning an optional error
is a much better way to go about it.

While fixing that, this also takes care of the new invariant that
blanks can only be on the head and tail of a sequence.  They aren't
allowed internally anymore--so no paths like foo//bar (though
/foo/bar/ is still legal).  This restriction has a few benefits,
but most notably it avoids contention where URLs like:

  http://example.com

...won't be interpreted as a 3-element PATH! with a CHAIN! in the
first slot, a BLANK! in the middle, and a TUPLE! in the third.

There had yet to be any meaningful uses of blanks internally to
paths or tuples.  If they wind up being needed, they could be
permitted as foo/_/bar ... but this would make blank's rendering
inconsistent.