Sort out Root_Heavy_Null and Root_Heavy_Void
There were plain blocks being created which were claiming to be
heavy null and heavy void.  However, in practice all they were doing was
to hold alive the arrays of a single quasi void and single quasi null.

Making them actual heavy nulls and voids would involve making them
Atom*, which is illegal in an API handle.  So accuracy on making them
Atoms would subvert the advantage of using an API handle to hold the
arrays alive without hardcoding references to them in the GC.

This clarifies the situation by putting the code more together, and also
changes to say that they're meta-heavy-null and meta-heavy-void, using
quasiforms so as to be API legal.