Introduce chokepoints for quasiform/antiform creation
There are emerging rules for quasiforms and antiforms...such as that
only certain types are allowed to be "isotopic" (word is, path isn't,
etc.)  But also that antiforms can't have bindings.  And only some
word quasiforms can be turned into antiforms.  This goes along with
other rules, like antiform null can't be put into API cells...and
unstable antiforms can't be put in API cells either.

This introduces some mechanisms to force all callers through a
common code path for producing quasiforms and antiforms, so that
checks can be done.  For the debug build, it turns the ANTIFORM_0 and
QUASIFORM_2 quote byte constants into a C++ object that can be used
for comparisons but not assignments--restricting the places that
assignments are done to a few places.