Get rid of some unused PARSE3 META-GROUP! code
Some speedbump in history had a moment where it seemed like a good
idea for specially annotated groups to represent a desire to not
splice insertions/changes.  (It was probably @(...) groups).  At
some point that got shuffled to ^(...) groups.  Some forgotten code
in PARSE3 still had that.

Because PARSE3 only exists for bootstrap compatibility, using a
^ form that the bootstrap executable can't make is clearly useless,
beyond the other uselessness.  This deletes that, and just does the
splicing by default for the feature--since it's what bootstrap
does and it's more general (you can "avoid splicing" by putting your
data in a single element block).