"Tweak" some all caps names
Because the functions that "higher level" code are supposed to call tend
so start with Init_Xxx(), I had some reticence in changing the "lower
level" macros and functions that did little (to no) validation to also
be called Init_Xxx() in a way that made it indistinguishable as to
whether it was something that privileged code should call or not.

Rather than capitalization, this goes with the idea of using a more
"privileged-seeming" word: Tweak.  This suggests that you're fiddling
with a part of a structure, and not initializing it in its entirely.