Terminal slash suppresses action execution
While some debate has been had over what the most useful meaning for
terminal slash is, it appears the best idea for it is to get the
action without running it.

This is a very commonly needed thing, and representing it with a
slash at the end offers the dual benefit of saying "this is something
dealing with function dispatch" (as slashes now are exclusively),
but also to throw up a kind of block between the function and any
potential arguments...to help reinforce that it doesn't take them.

    for-next: specialize for-skip/ [skip: 1]
    ;                         ---^
    ; slash helps show block is not argument

This looks much nicer than using get:

    for-next: specialize get $for-skip [skip: 1]

In particular, because GET may soon not be willing to return an
action unless you say GET:ACTIVE or similar.