Get rid of NIHIL function, NOTHING => value, not func
There was potential for confusion by having the states named NOTHING
and NIHIL be functions vs. their values.  This is because word
references cannot produce those values unless as the product of a
function.  But this leads to potential confusion, e.g. people
thinking that ^NOTHING or ^NIHIL would be the metaforms of those
values, as opposed to the meta of the function that produces them.

Using the literal form ~[]~ is fine for NIHIL and actually looks
better in many contexts, drawing attention to its unique nature.

The definition of NOTHING is changed to be actually an unset
variable.  This can't be used in its plain form via word access,
but ^NOTHING has meaning.