Organize Get_Var() routines, return trapped errors
The Get_Var() routines had become fairly organic, making it not clear
when you should call the higher level routines or lower level ones.
This makes sure the evaluator has separate entry points for words
and tuples and paths that are appropriate, and makes it clear when
vacant states may be returned.

In addition, it starts making it possible to trap errors from things
like variables not being bound.  This makes it possible to do things
like TRY GET of a variable, or to GET with an EXCEPT handler...vs.
needing to use RESCUE.  (Some cases still have to abruptly fail, due
to being in the middle of a process and representing an amount of
failure that you likely shouldn't absorb it and keep going.)