Rename Context* to VarList*
The VarList is a sequential structure with two parallel arrays, one
for keys and one for values.  It is now no longer the only way that
contexts are implemented--because Module uses the "sea of words"
by hanging the variables in chains off the symbol table.  But there
are plans now to probably subsume LET and USE structures into the
category of contexts.

What is currently called a "specifier" is slated to become context,
with it spanning these new classes.  So VarList is given a more
focused name.  (It still includes module contexts, which it should
not...so that will be sorted out as well.)