Speak in terms of signed integers
There was a pervasive tax on the codebase from converting between signed
and unsigned integer types.  This is considered to be a bad practice,
and even the choice of `size_t` to be unsigned is considered to be
a mistake by Bjarne Stroustrup:

  https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2019/p1428r0.pdf

Certainly it wasn't helping anything, in particular because nothing
stops signed values from being passed to unsigned slots or vice versa.
So there were tons of pointless casts.

This gets rid of those casts and paves the way for better checks.