Rethink UNLIMITED as optional pointer vs. -1
The idea of using a magic number that's a cast of a large negative
number is very error prone.  But Optional() cannot be used when 0
is a legal in-band value.

So this just goes and uses pass-by-pointer as the convention for
"limit", and if that pointer is null it means no limit.  This means
that the Optional() can be used, which enforces a check before use
and going through an `unwrap` function to use the value.