Speculative work: 1-based Index type
For a long time, I've wondered if it would be useful to create a
type that represented 1-based access.  It's a little hard to enforce
a type system that checks that, because having typecheck errors on
1-based values vs. 0-based values is a burden.  It can at least
serve a documentation purpose--and if the 1-based type is barred
from holding 0 unless it is specifically Optional() that may have
at least some limited benefit.

(Due to the fact that Optional(Index) becomes just `Index` if not
using DEBUG_CHECK_OPTIONALS, the assert that index isn't initialized
to 0 is not able to be enforced unless that switch is set.)