Restyle Optional() helpers to permit template specialization
The ability to say `unwrap optional` instead of `unwrap(optional)` is
a trick based on using a stream insertion operator (<<) to get that as
a macro that does `g_unwrap_helper << optional`.  This wasn't styled
in a way that let you make Optional template overloads, but this
changes that so << is a global function not a member function.