Change Maybe_Locate_Token() to use Trap convention
The locating token routine was returning a TOKEN_0 as the primary
return result when there was an error, and the error was filled in as
an output argument by pointer.  This is counter to the convention that
is starting to emerge of using Trap_XXX routines that return an
Optional() error.  This flips that around.