Reorganize scanner into clearer sections
An attempt to make Token a local variable in the Scanner_Executor()
instead of a property of the scanner Level revealed what was already
apparent--that the stackless reorganization of the scanner had gotten
out of control.

While rewriting the scanner completely would be a nice thing to have
the time to do, this does at least reorganize it into sections that
make more sense...and by removing the token from the Level it makes
the code more "functional" in the sense that fewer parts can be made
to accidentally depend on the last token seen.