Tighten LEX type
This makes it so that any accesses of the Lex_Map go through what
can be made a function in a checking mode, that ensures the input
is a single byte character.  (The previous code would cast whatever
you passed in to a byte, making it accept arbitrary arguments
which could conflate to bytes with the wrong meaning.)

There was also some sort convoluted logic for using the Lex_Map
to quickly turn a character into a hex digit, which was repeated
several times.  This captures that in a single inline function.