Rename Did_XXX Functions to Try_XXX
There was some idea that DID could be used to denote returning a
logic, while TRY would be used when you wanted to return a pointer
to a thing that may be null or not.

It just comes out looking silly.  It's okay for you to TRY something
that writes its output somewhere else dictated by a parameter, and
the result of the Try_XXX routine is a boolean.  Looks much better.