Make scanning helpers push to stack directly
At some point Ren-C changed the Rebol3 convention of scan helpers
pushing directly to the stack to taking the cell to write the
scanned data.  The idea was to generalize those routines to they
were more useful to places that didn't build values on the stack.

This turned out to be a bad idea, as it meant the scanner was
pushing a cell on the stack which may or may not get filled with
data in the case of an error.  That motivated some strange
attempts to "hook return" via a `return_NULL` that would erase
the cell, etc.

This commit goes back to the simpler protocol, but with modern
improvements using Option() to make the return convention
clearer, as well as saying the scan routines go _To_Stack() to
make it clearer.