Allow _.(void) to compose as _
The idea that the concept of being a sequence at all vaporizes if you
void out all the sequence slots is a strange idea (which is why it is
hidden under the /CONFLATE option of COMPOSE).  But if it's going to
work, then it should probably work with blanks as well as with any
other type--like WORD!.

But Trap_Check_Sequence_Element() will complain about the blank since
only special testing permits it at the head and tail. Subvert the
check when doing the conflation.