Solidify "Pairing" code
The special stub-sized type that holds two cells is called a "pairing",
but it didn't have a unique type.  It was just a Cell, that was
considered to be "paired" with another cell.  That concept doesn't
really have much merit compared to having a Pairing abstraction that
is derived from Node that you have to manually get the cells out of.

This adds that datatype to abstract it.  It also gets rid of some
lingering code pertaining to floating point pairs, which it is
telling that Red didn't bother with implementing.  There was a fairly
involved scanning routine that was used just for that feature.