Restyle code for natives at head, GET-WORD?/SET-WORD?
Some natives are used in typechecking for other natives.  When that
happens, the order in which they are loaded matters.  There was some
code in the %make-natives.r script which was hacked in to do this,
but it needed a better articulation...where the spec of what to do
was just the list of natives in order.  It had long since grown
long enough to be worth it to write the (very easy) code to replace
the hack.

Also adds SET-WORD? and GET-WORD? as natives, which they need to
be in upcoming commits.