Cut down on SET-WORD! and GET-WORD! usages
Because these will no longer be fundamental types (but instances of
CHAIN!) there are going to have to be new approaches to dealing with
places that had SET-WORD! or GET-WORD! mentions before.

This fixes up some number of those with equivalent code, in order to
make it clearer what the material changes of the "convert to chain"
commit actually are.