Eliminate "attachment binding", strict mode always
Originally Rebol would create variables for every word that it saw,
in case that word was ever used as a variable.  This led to an effect
that was kind of like "non-strict-mode" in JavaScript, where any
SET-WORD! embedded at any level in the code--where a LET or `<local>`
definition was forgotten--could write into global space.

R3-Alpha actually had a stricter policy for the LIB and SYS contexts,
but everywhere else was non-strict.

In Ren-C's model, there are not variables created in advance for
everything.  So to simulate non-strict mode required something called
"attachment binding".  It was bad and created some nebulous semantics,
so getting rid of it is the best idea.

This mitigates the impact by making the console add top-level set-words
in the expressions you type to the user context.