Outdent the scanner code
In a classic example of "we should be editing our code as a graph
database", this massive seeming commit is just outdenting the scanner
by one tab level.

Just makes it easier to fit more on the page and to borrow code
from the "modern" (slightly more modern) scanner, which is indented
at this level.