Start to rethink ScanState and TranscodeState
There's always been a necessity to break out scanning to some
information that's global for the scan and some that's specific
to a layer/level of scanning, but due to rewriting the scanner
being low priority it didn't get a lot of attention.

Because information gets out of sync it would be nice to be
referencing the fields directly, so S->begin and S->end might
be nicer than bp and ep that are constantly getting out of
sync.  This is a first phase in moving in that direction.