Constrain scope of start and end token in scanner
The start and end token locations in the scanner were part of a global
state, when it seemed the global state really only needed to track the
current furthest-scanned point.  But because the states were mixed up
with each other, it was easy to get confused and have them out of sync.

This now makes it so that the begin and end tokens are both generated
for each scanner level from the global transcode state's single
position, allowing that position to be more narrowly and clearly
managed in the code.