Make custom RAISE macro for the scanner executor
This makes it so that errors which are returned from the various scan
components can all be augmented automatically to show the line and file
being scanned.  Previously this was only done by e.g. Error_Syntax(),
which did isolate a token out, but may be more vague than if you had
the file and line information and an error more specific about what
it was that was going wrong.