Split S->begin and S->end from transcode state
This helps untangle some of the more puzzling behaviors that
came from the juggling of begin/end to mean too many things.
It has one pointer in the global state for where the next
data to consume is, and makes the S->begin and S->end state
for the token tied to a scan level only.

The bp and ep pointers are now removed from consideration
outside of individual switch cases that use them as a
convenience.