Terminal slash will get and specialize actions
This allows you now to use append/ as an alternative to :append, but
one that will fail if you try it on other values.

If you say append/dup/ then you will get a specialization.  Since the
chain format is now working as equivalent, you can even write append:dup/
for compatibility in bootstrap.