Commit to @(xxx) as "soft literal" format
This goes with what the plan has been for some time, bolstered by the
idea that refinements will be @:(xxx) instead of @(:xxx)

Getting this to work required trimming some of the no-longer-working
native combinator code for UPARSE, which still had lingering ideas of
using the REMAINDER system-multi-return proxying, which has now been
deleted.  All of that needs rewriting.