Run crypt tests, fix CHECKSUM leak
The crypt module's tests weren't being run (there needs to be a way
to detect which extensions are built and run their tests...)

For now this just adds them in, and that triggered an assert on a
leak that was in the CHECKSUM function.  So this fixes that, and also
brings the tests in line with the new conventions.