Embrace more "modernism" in the bootstrap shim
Originally I had some reluctance to making the code in the bootstrap
and import shim use things that looked like CHAIN! or VAR-WORD! etc.
But really, it has to be reliable to load these as paths...otherwise
things wouldn't work.

This goes ahead and embraces it more, and also the bootstrap EXE
was changed to use :BLAME instead of :WHERE with FAIL, so that's one
less difference to worry about in the code.