Eliminate illegal antiform use of ~done~
There was some runaround with ~done~ for evaluating in the bootstrap
build, which raised attention to the fact that the modern build was
actually creating an antiform of ~done~, which is not currently a
legal "keyword".

It turned out the evaluator wasn't going through the mechanics for
checking for illegal antiforms, and it was only being caught via
other ancillary mechanisms when you typed in the console.  So this
fixes that.

Rather than try and rig the bootstrap build up to somehow use a
tripwire, this just uses a quasi-word instead, which accomplishes
the goal of the sanity checks and is simpler.