Put all infix bits into action cells
The design of infix has gone through a lot of changes, but settled on
a form with 4 basic modes: no infix, tight infix, deferred infix, and
postponed infix.

These modes can be fit into a cell's header for frames, and it's best
if they are--since not doing so causes all sorts of headaches.

This not only fits the bits into the cell but also transfers those
bits into the action level so the action executor knows what infix mode
a function call was launched with.  This is exposing the convolutions
of some code and will hopefully enable it to be cleaned up.